Introduce VectorBaseDialogFragment
